Split extension up into multiple files
