+++ packages/vscode-extension/src/language-server/features/hover.ts
@@ -1,0 +1,11 @@
+import { Connection } from "vscode-languageserver/browser";
+import { DocumentInfo } from "../globals";
+
+export function addDocumentHoverSupport(
+    connection: Connection,
+    documentInfo: DocumentInfo,
+) {
+    connection.onHover((params) => {
+        return null;
+    });
+}
